Use operator.slug for URLs, not slugify
